requirements.txt
google-cloud-speech>=2.26.0
gTTS>=2.5.1
numpy>=1.24
scipy>=1.10
